@Green-Sky 2025-08-02
sd runs:

## `sd2 turbo q8_0 512x512`

### master
without flash attention:
```
  |==================================================| 8/8 - 6.48it/s
[INFO ] stable-diffusion.cpp:1822 - sampling completed, taking 1.34s
```

with flash attention:
```
  |==================================================| 8/8 - 9.38it/s
[INFO ] stable-diffusion.cpp:1822 - sampling completed, taking 0.96s
```

vae:
```
[INFO ] stable-diffusion.cpp:1843 - latent 1 decoded, taking 0.76s
```

### pr
without flash attention:
```
  |==================================================| 8/8 - 7.18it/s
[INFO ] stable-diffusion.cpp:1822 - sampling completed, taking 1.30s
```

with flash attention:
```
  |==================================================| 8/8 - 10.85it/s
[INFO ] stable-diffusion.cpp:1822 - sampling completed, taking 0.85s
```

vae:
```
[INFO ] stable-diffusion.cpp:1843 - latent 1 decoded, taking 0.47s
```


## `sd1 f16 512x768`

### master
without flash attention:
```
  |==================================================| 30/30 - 1.40it/s
[INFO ] stable-diffusion.cpp:1822 - sampling completed, taking 20.89s
```

with flash attention:
```
  |==================================================| 30/30 - 1.52it/s
[INFO ] stable-diffusion.cpp:1822 - sampling completed, taking 19.65s
```

vae:
```
[INFO ] stable-diffusion.cpp:1843 - latent 1 decoded, taking 1.20s
```

### pr
without flash attention:
```
  |==================================================| 30/30 - 1.57it/s
[INFO ] stable-diffusion.cpp:1822 - sampling completed, taking 19.16s
```

with flash attention:
```
  |==================================================| 30/30 - 1.65it/s
[INFO ] stable-diffusion.cpp:1822 - sampling completed, taking 18.15s
```

vae:
```
[INFO ] stable-diffusion.cpp:1843 - latent 1 decoded, taking 0.74s
```

---

@leejet working hard to widen the performance gap between vulkan and cuda again (: